improved validation
